+++ src/command_modules/azure-cli-profile/azure/cli/command_modules/profile/custom.py
@@ -92,26 +92,24 @@ def account_clear(cmd):
 
 # pylint: disable=inconsistent-return-statements
 def login(cmd, username=None, password=None, service_principal=None, tenant=None, allow_no_subscriptions=False,
-          identity=False, identity_port=None,
-          msi=False, msi_port=None):  # will remove msi_xxx in a future release
+          identity=False, identity_port=None):
     """Log in to access Azure subscriptions"""
     from adal.adal_error import AdalError
     import requests
 
     # quick argument usage check
-    if (any([password, service_principal, tenant, allow_no_subscriptions]) and
-            any([identity, msi])):
+    if (any([password, service_principal, tenant, allow_no_subscriptions]) and identity):
         raise CLIError("usage error: '--identity' is not applicable with other arguments")
 
-    if msi_port or identity_port:
-        logger.warning("'--msi-port/--identity-port' is no longer required to login using managed identity."
+    if identity_port:
+        logger.warning("'--identity-port' is no longer required to login using managed identity."
                        " This flag will be removed in a future release of CLI.")
 
     interactive = False
 
     profile = Profile(cli_ctx=cmd.cli_ctx, async_persist=False)
 
-    if identity or msi:
+    if identity:
         if in_cloud_console():
             return profile.find_subscriptions_in_cloud_console()
         return profile.find_subscriptions_in_vm_with_msi(username)