+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -4,6 +4,7 @@ Release History
 ===============
 0.1.31
 ++++++
+* (Breaking change): remove `assign-identity` which was tagged `deprecating` 2 releases ago
 * webapp: capture the unhandled exception if the appservice plan doesn't exist
 
 0.1.30